add comments, some cleanup
