+++ lib/pypechain/README.md
@@ -1,1 +1,16 @@
 Pypechain generates web3.py Contract classes for a given solidity contract abi.
+
+### Install
+
+See toplevel INSTALL.md
+
+### Usage
+
+From lib/pypecahin, run:
+
+```bash
+#  python pypechain/run_pypechain.py <ABI_FILE>          <OUT_FILE>
+❯❯ python pypechain/run_pypechain.py './abis/ERC20.json' './build/ERC20Contract.py'
+```
+
+Much of this is subject to change as more features are fleshed out.
