--- lib/pypechain/pypechain/example/callContract.py
@@ -1,6 +1,0 @@
-# try:
-#     # Instantiate contract
-#     erc20Contract = ERC20Contract(abi=contract_abi)
-#     fn = erc20Contract.functions.allowance("asdf", "asdf").call()
-# except FallbackNotFound:
-#     print("Fallback function not found. Continuing...")
