+++ lib/pypechain/pypechain/utilities/format.py
@@ -1,3 +1,4 @@
+"""Formatting utilities."""
 import keyword
